feat(protocol): add EnumField support + MESSAGE_NAME config
EnumField support (typed enums instead of int):
- Add enum definitions in protocol/enums/
- Generate typed enum classes (ChildType, DeviceType, etc.)
- Update struct generators to use enum types
- Update handlers and UI to use typed enums

MESSAGE_NAME configuration:
- serial8_protocol_config.py: add include_message_name=True

Generated code regenerated with protocol-codegen updates.

+++ src/handler/input/HandlerInputViewState.cpp
@@ -55,7 +55,7 @@ void HandlerInputViewState::sendViewState() {
     OC_LOG_DEBUG("[ViewState] Sending: view={}, selectorActive={}",
                  static_cast<int>(currentView), selectorActive);
 
-    protocol_.viewState(static_cast<uint8_t>(currentView), selectorActive);
+    protocol_.viewState(currentView, selectorActive);
 }
 
 }  // namespace bitwig::handler
